Generate and validate token functionalities implemented
