+++ Commons/src/main/java/org/example/Roles.java
@@ -1,0 +1,8 @@
+package org.example;
+
+public enum Roles {
+    BUYER_I, // view/create/send/edit PO
+    BUYER_II,  // view/pay INVOICE
+    SUPPLIER_I, // view/create/edit/send INVOICE, view approved PO
+    SUPPLIER_II // approve/reject PO
+}
