+++ Commons/src/main/java/org/example/errorhandling/ExceptionResponseDTO.java
@@ -1,0 +1,16 @@
+package org.example.errorhandling;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ExceptionResponseDTO {
+    String error;
+    Integer status;
+    String details;
+}
