+++ Commons/src/main/java/org/example/errorhandling/InvalidTokenException.java
@@ -1,0 +1,4 @@
+package org.example.errorhandling;
+
+public class InvalidTokenException extends RuntimeException{
+}
