+++ PurchaseOrder/pom.xml
@@ -37,14 +37,9 @@
             <artifactId>springdoc-openapi-ui</artifactId>
             <version>1.7.0</version>
         </dependency>
+
         <dependency>
             <groupId>org.example</groupId>
-            <artifactId>core</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.system.automation</groupId>
             <artifactId>Commons</artifactId>
             <version>0.0.1-SNAPSHOT</version>
             <scope>compile</scope>
