--- UserAuth/src/main/java/com/system/automation/business/errorhandling/ErrorMessages.java
@@ -1,6 +1,0 @@
-package com.system.automation.business.errorhandling;
-
-public class ErrorMessages {
-    public static final String ALREADY_EXISTING_USER = "Already existing user identified by ";
-    public static final String USER_NOT_FOUND = "User not found: ";
-}
