--- UserAuth/src/main/java/com/system/automation/persistence/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.system.automation.persistence.repository;
-
-import com.system.automation.persistence.collections.User;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserRepository extends MongoRepository<User, UUID> {
-    Optional<User> findByUsername(String username);
-}
