--- UserAuth/src/main/java/com/system/automation/presentation/utils/Password.java
@@ -1,13 +1,0 @@
-package com.system.automation.presentation.utils;
-
-
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-
-public class Password {
-    private static final BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder(15);
-
-    public static String encrypt(String password){
-        return bCryptPasswordEncoder.encode(password);
-    }
-}
