--- UserAuth/src/main/java/com/system/automation/presentation/view/LoginRequestDTO.java
@@ -1,15 +1,0 @@
-package com.system.automation.presentation.view;
-
-import lombok.*;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Setter
-@Getter
-public class LoginRequestDTO {
-    private String username;
-
-    private String password;
-}
