--- UserAuth/src/main/java/com/system/automation/presentation/view/LoginResponseDTO.java
@@ -1,13 +1,0 @@
-package com.system.automation.presentation.view;
-
-import lombok.*;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Setter
-@Getter
-public class LoginResponseDTO {
-    String jwsToken;
-}
