+++ UserAuth/src/main/java/org/example/business/errorhandling/ErrorMessages.java
@@ -1,0 +1,7 @@
+package org.example.business.errorhandling;
+
+public class ErrorMessages {
+    public static final String ALREADY_EXISTING_USER = "Already existing user identified by ";
+    public static final String USER_NOT_FOUND = "User not found: ";
+    public static final String INVALID_CREDENTIALS = "Invalid credentials";
+}
