+++ UserAuth/src/main/java/org/example/business/errorhandling/customexceptions/InvalidCredentialsException.java
@@ -1,0 +1,7 @@
+package org.example.business.errorhandling.customexceptions;
+
+public class InvalidCredentialsException extends RuntimeException{
+    public InvalidCredentialsException(String message) {
+        super(message);
+    }
+}
