+++ UserAuth/src/main/java/org/example/persistence/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package org.example.persistence.repository;
+
+import org.example.persistence.collections.User;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends MongoRepository<User, UUID> {
+    Optional<User> findByUsername(String username);
+}
