+++ UserAuth/src/main/java/org/example/presentation/view/LoginRequestDTO.java
@@ -1,0 +1,15 @@
+package org.example.presentation.view;
+
+import lombok.*;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+public class LoginRequestDTO {
+    private String username;
+
+    private String password;
+}
