+++ UserAuth/src/main/java/org/example/presentation/view/LoginResponseDTO.java
@@ -1,0 +1,13 @@
+package org.example.presentation.view;
+
+import lombok.*;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+public class LoginResponseDTO {
+    String jwsToken;
+}
