+++ UserAuth/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.data.mongodb.uri=mongodb://localhost:27017/?retryWrites=true&serverSelectionTimeoutMS=5000&connectTimeoutMS=10000
+spring.data.mongodb.database=APsystem
